+++ PlusClouds.Net/Models/CurrencyModel.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace PlusCloudsNet.Models
 {
     public class CurrencyModel
     {
-        public int Value { get; set; }
+        public Double Value { get; set; }
         public string Formated { get; set; }
         public string Currency { get; set; }
     }
