+++ PlusClouds.Net/Request/Vm/VmAddIpRequest.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace PlusCloudsNet.Request.Vm
 {
     public class VmAddIpRequest : VmJobRequest
     {
-        public int ForService { get; set; }
+        public Int64 ForService { get; set; }
         public dynamic BrmData { get; set; }
     }
 }
