+++ PlusClouds.Net/Response/Vm/VmJobResponse.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace PlusCloudsNet.Response.Vm
 {
     public class VmJobResponse : BaseResponse
@@ -14,8 +16,8 @@ namespace PlusCloudsNet.Response.Vm
 
     public class VmStatus
     {
-        public int Id { get; set; }
-        public int ServiceId { get; set; }
+        public Int64 Id { get; set; }
+        public Int64 ServiceId { get; set; }
         public string Status { get; set; }
     }
 }
